Add a replacing NaN with null version of save JSON function
Add a replacing NaN with null version of save JSON function. Currently, if there is any `nan` in the input files or during calculations within `simulariumio` using `numpy`, it might break the output JSON since JSON does not support `NaN` as a valid numerical value.

This new implemented `save_replacing_nan` saves the `TrajectoryData` to a `JSON` while replacing all `NaN` with `null` (which is supported by `JSON`). This should fix the error you get when uploading  the `.simularium` to the simularium website converted from `NERDSS` trajectory with `simulariumio`.